power: move power API to separate file
